+++ DelvUI/changelog.md
@@ -7,6 +7,8 @@ Fixes:
 - Fixed clicks not working on game windows when they are on top of unit frames or party frames.
 - Fixed unit frame background color and death indicator background color being linked together.
 - Fixed storm's eye bar for warrior not tracking the buff duration properly.
+- Fixed mana bars not being visible when previewing party frames.
+- Fixed health and mana values not being correctly displayed when previewing party frames.
 - Fixed mouse clicks sometimes not responding at all.
 
 # 0.3.1.2